Decorators and session checking are complete

+++ boom_digital/administrator/views.py
@@ -1,8 +1,11 @@
 from django.http import JsonResponse
-from django.shortcuts import redirect, render, get_object_or_404
+from django.shortcuts import redirect, render
 from django.views.decorators.http import require_POST
+from django.core.mail import send_mail
+from administrator.decorators import auth_admin
 
 from administrator.models import Admin_users, Brand, Category, Customer, Employee
+from boom_digital import settings
 from customer.models import Order
 from staff.models import Product
 
@@ -11,21 +14,22 @@ from staff.models import Product
 def loginfun(request) :
     msg = ''
     if request.method == 'POST' :
-        username = request.POST['username']
-        password = request.POST['password']
+        username = request.POST.get('username')
+        password = request.POST.get('password')
 
         try :
             admin = Admin_users.objects.get(
                  username = username,
                  password = password
             )
-            request.session['admin_sessionID'] = admin.id
+            request.session["admin_sessionID"] = admin.id
             return redirect('admin:dashboard')
         except :
             msg = 'invalid username or password !'
 
     return render(request, 'administrator/login.html', {'errorMessage' : msg })
 
+@auth_admin
 def dashboardfun(request):
     productList = Product.objects.all()
     proCount = productList.count()
@@ -37,7 +41,7 @@ def dashboardfun(request):
     orderCount = orderList.count()
     return render(request, 'administrator/dashboard.html',{'productCount':proCount,'customerCount':cusCount,'employeeCount':empCount,'orderCount':orderCount})
 
-   
+@auth_admin   
 def newCategoryfun(request):
     msg = ""
     if request.method == 'POST':
@@ -54,32 +58,39 @@ def newCategoryfun(request):
         msg = "Category added successfully"
     return render(request, 'administrator/categoryReg.html',{'successMessage':msg})
 
+@auth_admin   
 def categoryfun(request):
     categories = Category.objects.all()
     count = categories.count()
     return render(request, 'administrator/categories.html',{'categories':categories, 'category_count':count})
 
 def delCategoryfun(request,Category_id):
-    delItem = Category.objects.get(
-        id = Category_id,
-    )
-    delItem.delete()
-    return redirect('admin:category')
+    if "admin_sessionID" in request.session:
+        delItem = Category.objects.get(
+            id = Category_id,
+        )
+        delItem.delete()
+        return redirect('admin:category')
+    else:
+        return redirect("admin:login")
 
 def viewCategoryfun(request,Category_id):
-    editItem = Category.objects.get(
-        id = Category_id,
-    )
-    if request.method == "POST":
-        editItem.name = request.POST.get('category_name')
-        editItem.description = request.POST.get('description')
+    if "admin_sessionID" in request.session:
+        editItem = Category.objects.get(
+            id = Category_id,
+        )
+        if request.method == "POST":
+            editItem.name = request.POST.get('category_name')
+            editItem.description = request.POST.get('description')
 
-    if 'logo' in request.FILES:
-        editItem.logo = request.FILES.get('logo')
-    editItem.save()
-    return render(request, 'administrator/viewCategory.html',{'item':editItem})
+        if 'logo' in request.FILES:
+            editItem.logo = request.FILES.get('logo')
+        editItem.save()
+        return render(request, 'administrator/viewCategory.html',{'item':editItem})
+    else:
+        return redirect("admin:login")
 
-
+@auth_admin   
 def newBrandfun(request):
     categories = Category.objects.all()
     msg = ""
@@ -99,54 +110,86 @@ def newBrandfun(request):
         msg = "Brand added successfully"
     return render(request, 'administrator/brandReg.html',{'categories':categories, 'successMessage':msg})
 
+@auth_admin
 def brandfun(request):
     brands = Brand.objects.all()
     count = brands.count()
     return render(request, 'administrator/brands.html',{'brands':brands, 'brand_count':count})
 
 def delbrandfun(request,brand_id):
-    delItem = Brand.objects.get(
-        id = brand_id,
-    )
-    delItem.delete()
-    return redirect('admin:brand')
+    if "admin_sessionID" in request.session:
+
+        delItem = Brand.objects.get(
+            id = brand_id,
+        )
+        delItem.delete()
+        return redirect('admin:brand')
+    else:
+        return redirect("admin:login")
+
 
 def viewBrandfun(request,brand_id):
-    editItem = Brand.objects.get(
-        id = brand_id,
-    )
-    categories = Category.objects.exclude(
-        id = editItem.category_id
-    )#for exclude the default brand's category
+    if "admin_sessionID" in request.session:
+        editItem = Brand.objects.get(
+            id = brand_id,
+        )
+        categories = Category.objects.exclude(
+            id = editItem.category_id
+        )#for exclude the default brand's category
 
-    if request.method == "POST":
-        editItem.name = request.POST.get('name')
-        editItem.description = request.POST.get('description')
+        if request.method == "POST":
+            editItem.name = request.POST.get('name')
+            editItem.description = request.POST.get('description')
 
-        category_id = request.POST.get('category')
-        category_instance = Category.objects.get(id=category_id)
-        editItem.category = category_instance
-
-
-    if 'logo' in request.FILES:
-        editItem.logo = request.FILES.get('logo')
-    editItem.save()
-    return render(request, 'administrator/viewBrand.html',{'item':editItem, 'categories':categories})
+            category_id = request.POST.get('category')
+            category_instance = Category.objects.get(id=category_id)
+            editItem.category = category_instance
 
 
+        if 'logo' in request.FILES:
+            editItem.logo = request.FILES.get('logo')
+        editItem.save()
+        return render(request, 'administrator/viewBrand.html',{'item':editItem, 'categories':categories})
+    else:
+        return redirect("admin:login")
+
+@auth_admin
 def newEmployeefun(request):
     msg = ""
     if request.method == 'POST':
+        first_name = request.POST.get('first_name')
+        second_name = request.POST.get('last_name')
+        email = request.POST.get('email')
+        phone = request.POST.get('number')
+        gender = request.POST.get('gender')
+        position = request.POST.get('position')
+        hired_date = request.POST.get('hired_date')
+        qualification = request.POST.get('qualification')
+        profile_img = request.FILES.get('profile_img')
+        # Generate a password
+        password = first_name.lower() + '@boom'
+        message = 'Hai Your username is ' + str(email) + 'and temporary password is ' + password
+        
+        # Create a new user with the generated password
         newEmoloyee = Employee (
-            first_name = request.POST.get('first_name'),
-            second_name = request.POST.get('last_name'),
-            email = request.POST.get('email'),
-            phone = request.POST.get('number'),
-            gender = request.POST.get('gender'),
-            position = request.POST.get('position'),
-            hired_date = request.POST.get('hired_date'),
-            qualification = request.POST.get('qualification'),
-            profile_img = request.FILES.get('profile_img'),
+            first_name = first_name,
+            second_name = second_name,
+            email = email,
+            password = password,
+            phone =phone,
+            gender = gender,
+            position = position,
+            hired_date = hired_date,
+            qualification =qualification,
+            profile_img = profile_img,
+        )
+        # Send email   
+        send_mail(
+            'Your New Account in Boom Digital Hub',
+            message,
+            settings.EMAIL_HOST_USER,
+            [request.POST.get('email')],
+            fail_silently = False
         )
         newEmoloyee.save()
         msg = "Registration completed" 
@@ -159,36 +202,47 @@ def employee_fun(request):
     return render(request, 'administrator/employeeList.html',{'employees':employees, 'employee_count':count})
 
 def delEmployeefun(request,employee_id):
-    delEmployee = Employee.objects.get(
-        id = employee_id,
-    )
-    delEmployee.delete()
-    return redirect('admin:employeeList')
+    if "admin_sessionID" in request.session:
+        delEmployee = Employee.objects.get(
+            id = employee_id,
+        )
+        delEmployee.delete()
+        return redirect('admin:employeeList')
+    else:
+        return redirect("admin:login")
+
 
 def employeeDetailsfun(request,employee_id):
-    editemployee = Employee.objects.get(
-        id = employee_id,
-    )
+    if "admin_sessionID" in request.session:
+        editemployee = Employee.objects.get(
+            id = employee_id,
+        )
 
-    if request.method == "POST":
-        editemployee.first_name = request.POST.get('first_name')
-        editemployee.second_name = request.POST.get('last_name')
-        editemployee.email = request.POST.get('email')
-        editemployee.phone = request.POST.get('number')
-        editemployee.gender = request.POST.get('gender')
-        editemployee.position = request.POST.get('position')
-        editemployee.qualification = request.POST.get('qualification')
-    if 'profile_img' in request.FILES:
-        editemployee.profile_img = request.FILES.get('profile_img')    
-    editemployee.save()
+        if request.method == "POST":
+            editemployee.first_name = request.POST.get('first_name')
+            editemployee.second_name = request.POST.get('last_name')
+            editemployee.email = request.POST.get('email')
+            editemployee.phone = request.POST.get('number')
+            editemployee.gender = request.POST.get('gender')
+            editemployee.position = request.POST.get('position')
+            editemployee.qualification = request.POST.get('qualification')
+        if 'profile_img' in request.FILES:
+            editemployee.profile_img = request.FILES.get('profile_img')    
+        editemployee.save()
 
-    return render(request, 'administrator/employeeDetails.html',{'editemployee':editemployee})
+        return render(request, 'administrator/employeeDetails.html',{'editemployee':editemployee})
 
+    else:
+        return redirect("admin:login")
+    
+
+@auth_admin
 def customerfun(request):
     customerList = Customer.objects.all()
     count = customerList.count()
     return render(request, 'administrator/customerList.html',{'customers':customerList,'customerCount':count})
 
+@auth_admin
 def stockfun(request):
     productList = Product.objects.all()
     productCount = productList.count()
@@ -197,6 +251,10 @@ def stockfun(request):
     total_current_stock = sum(product.current_stock for product in productList)
     return render(request, 'administrator/stockDetails.html',{'products':productList,'count':productCount, 'total_stock':total_current_stock})
 
+def logoutfun(request):
+    del request.session["admin_sessionID"]
+    request.session.flush()
+    return redirect("admin:login")  # Redirect to the common home page after logout
 
 def notFoundfun(request):
     return render(request, 'administrator/notFound.html')
